Bug 7685 - ask user beforehand
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@7881 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/specify/tasks/subpane/security/NavigationTreeContextMenuMgr.java
@@ -19,6 +19,8 @@
 */
 package edu.ku.brc.specify.tasks.subpane.security;
 
+import static edu.ku.brc.ui.UIRegistry.getResourceString;
+
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.awt.event.MouseAdapter;
@@ -41,7 +43,7 @@ import edu.ku.brc.dbsupport.DataProviderSessionIFace;
 import edu.ku.brc.specify.datamodel.Collection;
 import edu.ku.brc.specify.datamodel.SpPrincipal;
 import edu.ku.brc.specify.datamodel.SpecifyUser;
-import edu.ku.brc.ui.UIRegistry;
+import edu.ku.brc.ui.UIHelper;
 
 /**
  * An instance of this class manages the creation of the context (pop-up or right-click) menu that
@@ -115,11 +117,17 @@ public class NavigationTreeContextMenuMgr extends MouseAdapter implements TreeSe
             {
                 if (getTreeMgr().canDeleteUser(lastClickComp))
                 {
-                    getTreeMgr().deleteUser(lastClickComp);
+                    if (UIHelper.promptForAction("DELETE", "CANCEL", "SEC_ASKDELUSR_TITLE", getResourceString("SEC_ASKDELUSR")))
+                    {
+                        getTreeMgr().deleteUser(lastClickComp);
+                    }
                     
                 } else if (getTreeMgr().canRemoveUserFromGroup(lastClickComp))
                 {
-                    getTreeMgr().removeUserFromGroup(lastClickComp);
+                    if (UIHelper.promptForAction("REMOVE", "CANCEL", "SEC_ASKRMUSRGRP_TITLE", getResourceString("SEC_ASKRMUSRGRP")))
+                    {
+                        getTreeMgr().removeUserFromGroup(lastClickComp);
+                    }
                 }
                 lastClickComp = null;
                 updateBtnUI();
@@ -188,7 +196,7 @@ public class NavigationTreeContextMenuMgr extends MouseAdapter implements TreeSe
                         toolTip = "SEC_RM_USR_GRP";
                     }
                     delUserBtn.setEnabled(canDelUser || canRemUser);
-                    delUserBtn.setToolTipText(StringUtils.isNotEmpty(toolTip) ? UIRegistry.getResourceString(toolTip) : null);
+                    delUserBtn.setToolTipText(StringUtils.isNotEmpty(toolTip) ? getResourceString(toolTip) : null);
                 } else
                 {
                     delUserBtn.setEnabled(false);
